Fix CUDA and ROCm host utility stream bindings
