+++ rtp_llm/cpp/pybind/BUILD
@@ -63,7 +63,11 @@ cc_library(
         "@havenask//aios/autil:base64",
         "@havenask//aios/autil:zlib",
     ] + torch_deps() + select_py_bindings() + select({
-        "@//:using_cuda12": ["//rtp_llm/models_py/bindings/cuda/ops:gpu_base"],
+        "@//:using_cuda12": [
+            "//rtp_llm/cpp/utils:device_pin",
+            "//rtp_llm/models_py/bindings/cuda:cuda_host_utils",
+            "//rtp_llm/models_py/bindings/cuda/ops:gpu_base",
+        ],
         "@//:using_rocm": ["//rtp_llm/models_py/bindings/core:exec_ctx_rocm"],
         "//conditions:default": [],
     }) + no_block_copy_link_deps(),
